+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file. This projec
 ## Unreleased
 
 ### Added
-- New collection method `implode`.
+- New collection methods `implode`, `fill` and `fillObject`.
 - New standard iterator methods `take` and `tap`.
 - Collection can now be json encoded (implements `JsonSerializable`).
 - Collection can now be casted to a string.